# Conflicts:
#	src/main/java/logic/Astre.java
#	src/main/java/logic/Planete.java
#	src/main/java/logic/Simulation.java
#	src/main/java/view/PlaneteApp.java

+++ src/main/java/logic/CreateurPlanete.java
@@ -1,9 +1,11 @@
 package logic;
 
+import javafx.scene.paint.Color;
+
 public class CreateurPlanete extends CreateurAstre{
 
     @Override
-    public Planete factory(String nom, float taille, float masse, int pX, int pY){
+    public Planete factory(String nom, int taille, int masse, int pX, int pY){
         return new Planete(nom, taille, masse, pX, pY);
     }
 }
